docs(Toast): update Toast mobile demo
