+++ src/toast/__test__/demo.test.jsx
@@ -7,7 +7,9 @@ import iconVue from '@/toast/demos/icon.vue';
 import iconTextVue from '@/toast/demos/iconText.vue';
 import maskVue from '@/toast/demos/mask.vue';
 import mobileVue from '@/toast/demos/mobile.vue';
+import overlayVue from '@/toast/demos/overlay.vue';
 import positionVue from '@/toast/demos/position.vue';
+import preventScrollThroughVue from '@/toast/demos/preventScrollThrough.vue';
 import textVue from '@/toast/demos/text.vue';
 
 const mapper = {
@@ -15,7 +17,9 @@ const mapper = {
   iconTextVue,
   maskVue,
   mobileVue,
+  overlayVue,
   positionVue,
+  preventScrollThroughVue,
   textVue,
 };
